# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShaadiHub Wedding Marketplace API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VendorCategory:
      type: string
      enum:
        - Marriage Hall
        - Catering
        - Photography
        - Videography
        - Decoration
        - Car Rental
        - Event Planner
        - Makeup & Beauty
...
    VendorImage:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        imageUrl:
          type: string
        publicId:
          type: string
      required:
        - id
        - vendorId
        - imageUrl
        - publicId
    Package:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        features:
          type: array
          items:
            type: string
        isPopular:
          type: boolean
      required:
        - id
        - vendorId
        - name
        - price
        - features
        - isPopular
    Vendor:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        category:
          $ref: "#/components/schemas/VendorCategory"
        title:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        area:
          type: string
          nullable: true
        city:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        priceRangeMin:
          type: number
          nullable: true
        priceRangeMax:
          type: number
          nullable: true
        capacity:
          type: integer
          nullable: true
        whatsapp:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        verified:
          type: boolean
        isFeatured:
          type: boolean
        isSuspended:
          type: boolean
        rating:
          type: number
        totalReviews:
          type: integer
        profileViews:
          type: integer
        createdAt:
          type: string
          format: date-time
        images:
          type: array
          items:
            $ref: "#/components/schemas/VendorImage"
        packages:
          type: array
          items:
            $ref: "#/components/schemas/Package"
      required:
        - id
        - userId
        - category
        - title
        - city
        - verified
        - isFeatured
        - isSuspended
        - rating
        - totalReviews
        - profileViews
        - createdAt
        - images
        - packages
...
    VendorListResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
        vendors:
          type: array
          items:
            $ref: "#/components/schemas/Vendor"
      required:
        - total
        - page
        - pages
        - vendors
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: area
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
        - name: capacity
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Vendor list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorListResponse"

    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create vendor listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  $ref: "#/components/schemas/VendorCategory"
                title:
                  type: string
                description:
                  type: string
                location:
                  type: string
                area:
                  type: string
                city:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                priceRangeMin:
                  type: number
                priceRangeMax:
                  type: number
                capacity:
                  type: integer
                whatsapp:
                  type: string
              required:
                - category
                - title
      responses:
        "201":
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/me/profile:
    get:
      operationId: getMyVendorProfile
      tags: [vendors]
      summary: Get logged-in vendor profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Vendor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

    put:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  $ref: "#/components/schemas/VendorCategory"
                title:
                  type: string
                description:
                  type: string
                location:
                  type: string
                area:
                  type: string
                city:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                priceRangeMin:
                  type: number
                priceRangeMax:
                  type: number
                capacity:
                  type: integer
                whatsapp:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}/images:
    post:
      operationId: uploadVendorImage
      tags: [vendors]
      summary: Upload vendor image
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorImage"
  /vendors/{id}/images/{imageId}:
    delete:
      operationId: deleteVendorImage
      tags: [vendors]
      summary: Delete vendor image
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /vendors/{id}/packages:
    post:
      operationId: createPackage
      tags: [vendors]
      summary: Add pricing package
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                features:
                  type: array
                  items:
                    type: string
                isPopular:
                  type: boolean
              required:
                - name
                - price
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /vendors/{id}/packages/{pkgId}:
    put:
      operationId: updatePackage
      tags: [vendors]
      summary: Update package
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: pkgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                features:
                  type: array
                  items:
                    type: string
                isPopular:
                  type: boolean
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"

    delete:
      operationId: deletePackage
      tags: [vendors]
      summary: Delete package
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: pkgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # USERS
...